+++ src/vmware/CHANGELOG.md
@@ -1,8 +1,7 @@
 # Release History
 
 ## 3.0.0 (2021-06)
-This version has **breaking changes**:
-- `az vmware datastore create` has been removed. Please use `az vmware datastore netapp-volume create` or `az vmware datastore disk-pool-volume create` instead.
+- [BREAKING CHANGE] `az vmware datastore create` has been removed. Please use `az vmware datastore netapp-volume create` or `az vmware datastore disk-pool-volume create` instead.
 
 Other changes:
 - update to AVS 2021-06-01 API