+++ src/vmware/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Release History
 
-## 2.1.0 (TBD)
-- add az vmware private-cloud rotate-vcenter-password
-- add az vmware private-cloud rotate-nsxt-password
+## 3.0.0 (2021-06)
+- [BREAKING CHANGE] `az vmware datastore create` has been removed. Please use `az vmware datastore netapp-volume create` or `az vmware datastore disk-pool-volume create` instead.
+
+Other changes:
+- Update to AVS 2021-06-01 API
+- Add `az vmware private-cloud rotate-vcenter-password`
+- Add `az vmware private-cloud rotate-nsxt-password`
+- Add `az vmware addon` command group
+- Add `az vmware addon hcx` command group to manage a HCX addon
+- Add `az vmware addon srm` command group to manage a Site Recovery Manager (SRM) addon
+- Add `az vmware addon vr` command group to manage a vSphere Replication (VR) addon
 
 ## 2.0.1 (2021-02)
 - Update the minimum az cli version to 2.11.0 [#3045](https://github.com/Azure/azure-cli-extensions/issues/3045)